Moved docstring operation description to ViewInspector
