+++ rest_framework/schemas/inspectors.py
@@ -3,9 +3,13 @@ inspectors.py   # Per-endpoint view introspection
 
 See schemas.__init__.py for package overview.
 """
+import re
 from weakref import WeakKeyDictionary
 
+from django.utils.encoding import smart_text
+
 from rest_framework.settings import api_settings
+from rest_framework.utils import formatting
 
 
 class ViewInspector:
@@ -14,6 +18,9 @@ class ViewInspector:
 
     Provide subclass for per-view schema generation
     """
+
+    # Used in _get_description_section()
+    header_regex = re.compile('^[a-zA-Z][0-9A-Za-z_]*:')
 
     def __init__(self):
         self.instance_schemas = WeakKeyDictionary()
@@ -61,6 +68,45 @@ class ViewInspector:
     @view.deleter
     def view(self):
         self._view = None
+
+    def get_description(self, path, method):
+        """
+        Determine a link description.
+
+        This will be based on the method docstring if one exists,
+        or else the class docstring.
+        """
+        view = self.view
+
+        method_name = getattr(view, 'action', method.lower())
+        method_docstring = getattr(view, method_name, None).__doc__
+        if method_docstring:
+            # An explicit docstring on the method or action.
+            return self._get_description_section(view, method.lower(), formatting.dedent(smart_text(method_docstring)))
+        else:
+            return self._get_description_section(view, getattr(view, 'action', method.lower()),
+                                                 view.get_view_description())
+
+    def _get_description_section(self, view, header, description):
+        lines = [line for line in description.splitlines()]
+        current_section = ''
+        sections = {'': ''}
+
+        for line in lines:
+            if self.header_regex.match(line):
+                current_section, separator, lead = line.partition(':')
+                sections[current_section] = lead.strip()
+            else:
+                sections[current_section] += '\n' + line
+
+        # TODO: SCHEMA_COERCE_METHOD_NAMES appears here and in `SchemaGenerator.get_keys`
+        coerce_method_names = api_settings.SCHEMA_COERCE_METHOD_NAMES
+        if header in sections:
+            return sections[header].strip()
+        if header in coerce_method_names:
+            if coerce_method_names[header] in sections:
+                return sections[coerce_method_names[header]].strip()
+        return sections[''].strip()
 
 
 class DefaultSchema(ViewInspector):
